fix bugs

+++ calculation2.java
@@ -1,11 +1,11 @@
 public class calculation2 {
-	public int function2(int n){
-		int res=1;
+	public double function2(int n){
+		double res=1;
 		int i = 1;
 		while(i<=n){
-		       res=i*res;
-		       i=i+1;
-		       }
+		       res = i * res;
+		       i = i + 1;
+		}
 		return res;
 	}
 }
